Added north triggers for Vermilion Suicune scene
If the player waits to do suicune scenes 1 and 2 until after first reaching vermilion, they previously had to walk past a scriptless eusine and suicune to trigger the event at the south of the pier. Now, they can trigger the event from the north side of the pier so they shouldn't be able to interact with the scriptless npcs.